rename: 공통 커스텀 훅 폴더 위치 변경
